+++ src/main/java/uni/backend/service/HomeService.java
@@ -14,34 +14,10 @@ import uni.backend.domain.Role;
 import uni.backend.domain.dto.HomeDataResponse;
 import uni.backend.domain.dto.HomeProfileResponse;
 import uni.backend.repository.ProfileRepository;
+import uni.backend.util.MainCategoryMap;
 
 @Service
 public class HomeService {
-
-
-    private static final Map<String, String> HASHTAG_TRANSLATION_MAP = Map.ofEntries(
-        Map.entry("trip", "여행"),
-        Map.entry("旅行", "여행"),
-        Map.entry("administration", "행정"),
-        Map.entry("行政", "행정"),
-        Map.entry("realty", "부동산"),
-        Map.entry("房地产", "부동산"),
-        Map.entry("banking", "은행"),
-        Map.entry("银行", "은행"),
-        Map.entry("mobile", "휴대폰"),
-        Map.entry("通讯", "휴대폰"),
-        Map.entry("language exchange", "언어교환"),
-        Map.entry("语言交换", "언어교환"),
-        Map.entry("college life", "대학 생활"),
-        Map.entry("大学生活", "대학 생활"),
-        Map.entry("gastroventure", "맛집"),
-        Map.entry("美食游", "맛집"),
-        Map.entry("game", "게임"),
-        Map.entry("游戏", "게임"),
-        Map.entry("shopping", "쇼핑"),
-        Map.entry("购物", "쇼핑")
-    );
-
 
     private final ProfileRepository profileRepository;
 
@@ -92,7 +68,7 @@ public class HomeService {
         for (int i = 0; i < hashtags.size(); i++) {
             String originalTag = hashtags.get(i);
             String normalizedTag = originalTag.toLowerCase();
-            String koreanTag = HASHTAG_TRANSLATION_MAP.getOrDefault(normalizedTag,
+            String koreanTag = MainCategoryMap.HASHTAG_TRANSLATION_MAP.getOrDefault(normalizedTag,
                 originalTag); // 매핑되지 않으면 원본 유지
             hashtags.set(i, koreanTag);
         }
