+++ src/main/java/uni/backend/util/MainCategoryMap.java
@@ -1,0 +1,29 @@
+package uni.backend.util;
+
+import java.util.Map;
+
+public class MainCategoryMap {
+
+    public static final Map<String, String> HASHTAG_TRANSLATION_MAP = Map.ofEntries(
+        Map.entry("trip", "여행"),
+        Map.entry("旅行", "여행"),
+        Map.entry("administration", "행정"),
+        Map.entry("行政", "행정"),
+        Map.entry("realty", "부동산"),
+        Map.entry("房地产", "부동산"),
+        Map.entry("banking", "은행"),
+        Map.entry("银行", "은행"),
+        Map.entry("mobile", "휴대폰"),
+        Map.entry("通讯", "휴대폰"),
+        Map.entry("language exchange", "언어교환"),
+        Map.entry("语言交换", "언어교환"),
+        Map.entry("college life", "대학 생활"),
+        Map.entry("大学生活", "대학 생활"),
+        Map.entry("gastroventure", "맛집"),
+        Map.entry("美食游", "맛집"),
+        Map.entry("game", "게임"),
+        Map.entry("游戏", "게임"),
+        Map.entry("shopping", "쇼핑"),
+        Map.entry("购物", "쇼핑")
+    );
+}
